Server/requirements.txt
fastapi
uvicorn
Pillow
httpx